app/requirements.txt
Flask
authlib
pyOpenSSL
requests
